requirements.txt
# Core Framework
Flask==3.1.0
Werkzeug==3.1.3

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Excel/Reports
openpyxl==3.1.5
pandas==2.2.3
XlsxWriter==3.2.0

# HTTP/API
requests==2.32.3
urllib3==2.2.3

# Configuration
PyYAML==6.0.2
python-dotenv==1.0.1

# Scheduling
APScheduler==3.10.4

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Logging
colorlog==6.9.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0

# Development
black==24.10.0
flake8==7.1.1
isort==5.13.2
